Update information about installing with GPUs (#130)
I think we have discovered the cause of the mysterious hangs when you use CUDA! This PR documents it, as a short term "fix". In the future, I want to precompile more compute capabilities and/or make it print out status information when it's compiling.

Also I removed the MacOS GPU info because we don't provide MacOS CUDA builds any more.