Remove Xcode Playground documentation (#238)
The present `Usage.md` instructs Mac users download a toolchain and then to use it in an Xcode Playground. However:

> Swift for Tensorflow [does not officially support Xcode Playgrounds](https://github.com/tensorflow/swift/issues/206#issuecomment-510416213), and use there is [currently broken](https://bugs.swift.org/browse/TF-500).

This PR replaces the Xcode Playground instructions with that note, and it updates the remaining Mac usage instructions to cover the supported and working method of building a command line app.