Allow selectively skipping certain groups of tests on Linux (#244)
This is necessary for Google3 import where Quote and SIL are
imported as two different libraries.

By default, both Quote and SIL tests are running. If necessary, either
group of tests can be skipped via:

    swift test -Xswiftc -D -Xswiftc SKIP_QUOTE_TESTS
    swift test -Xswiftc -D -Xswiftc SKIP_SIL_TESTS

Incantations for Bazel/Blaze are irrelevant to this repository,
so I won't be including them here.

These incantations only work on Linux.